Reduce attack surface for Docker image
The current `.dockerignore` file is designed as deny-list. Representing
an allow-list reduced the overall amount of files that are sent to the
Docker build daemon and allows for easier and more fine-grained
specification of the copied files.

In example, `alerts.hpilo.yml`, the `grafana_dashboard.json` and the
whole `.github` folder shouldn't be part of the resulting image. This is
currently the case.

Signed-off-by: justusbunsi <61625851+justusbunsi@users.noreply.github.com>